# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_azzimoda_subscriberest_internal_model.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  internal_handler.CreateSubscriptionRequest:
    properties:
      end_date:
        example: 2026-12
        type: string
      price:
        example: 999
        minimum: 0
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 2026-01
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  internal_handler.CreateSubscriptionResponse:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      message:
        example: created
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        example: invalid subscription ID format, expected UUID
        type: string
    type: object
  internal_handler.GetStatsResponse:
    properties:
      meta:
        $ref: '#/definitions/internal_handler.StatsMeta'
      result:
        example: 2997
        type: integer
    type: object
  internal_handler.ListSubscriptionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_azzimoda_subscriberest_internal_model.Subscription'
        type: array
      meta:
        $ref: '#/definitions/internal_handler.PaginationMeta'
    type: object
  internal_handler.PaginationMeta:
    properties:
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 10
        type: integer
    type: object
  internal_handler.StatsMeta:
    properties:
      end_date:
        example: 2026-06
        type: string
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 2026-01
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  internal_handler.UpdateSubscriptionRequest:
    properties:
      end_date:
        example: 2026-12
        type: string
      price:
        example: 999
        minimum: 0
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 2026-01
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления подписками пользователей
  title: SubscribeREST API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Returns a paginated list of all subscriptions
      parameters:
      - default: 10
        description: Number of rows per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.ListSubscriptionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription for a user
      parameters:
      - description: Subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.CreateSubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Create a subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete a subscription by ID
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Delete a subscription
      tags:
      - subscriptions
    get:
      description: Returns a subscription by its UUID
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_azzimoda_subscriberest_internal_model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update an existing subscription by ID. All fields are optional
        — only provided fields will be updated.
      parameters:
      - description: Subscription ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_azzimoda_subscriberest_internal_model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Update a subscription
      tags:
      - subscriptions
  /subscriptions/stats:
    get:
      description: Calculate total cost of subscriptions for a user within a date
        range, optionally filtered by service name
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Filter by service name (partial match)
        in: query
        name: service_name
        type: string
      - description: Start date (YYYY-MM)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.GetStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Get subscription stats
      tags:
      - subscriptions
swagger: "2.0"